Use clouds.yaml instead of openrc or openrc.yaml
clouds.yaml is a yaml-based config format that is used by Ansible's
OpenStack modules and python-openstackclient. Go ahead and use it so
that convergence via kolla with upstream ansible openstack modules as
well as use of python-openstackclient can be consistent.

Also, remove the tenant_name setting, it's not needed by the ansible
modules or python-openstackclient. AND - project_domain_name and
user_domain_name should be used as the variables since the setting is
coming from keystone.domain_name.